Make the v1 training choose precision from the hardware
