+++ test/test_util.py
@@ -2,6 +2,7 @@ import os
 import unittest
 from shutil import rmtree
 
+import pytest
 import numpy as np
 import requests
 import torch
@@ -806,6 +807,27 @@ class TestEmbeddingBackend(unittest.TestCase):
             self.assertEqual(f.attrs["model_name"], "vit_t")
             self.assertEqual(list(f.attrs["original_size"]), [512, 512])
             self.assertIsNone(f.attrs["tile_shape"])
+
+
+@pytest.mark.parametrize(
+    "capability, expected", [((9, 0), torch.bfloat16), ((8, 0), torch.bfloat16), ((7, 5), None), ((6, 1), None)]
+)
+def test_the_training_precision_skips_the_float16_tier(monkeypatch, capability, expected):
+    """Inference runs float16 on Volta and Turing. Training stays in fp32 there, because it uses bfloat16."""
+    from micro_sam.util import training_autocast_dtype
+
+    monkeypatch.setattr(torch.cuda, "is_available", lambda: True)
+    monkeypatch.setattr(torch.cuda, "get_device_capability", lambda device: capability)
+
+    assert training_autocast_dtype(torch.device("cuda")) is expected
+
+
+@pytest.mark.parametrize("device_type", ("cpu", "mps"))
+def test_the_training_precision_is_fp32_outside_cuda(device_type):
+    """bfloat16 is emulated on these devices, so autocast makes training slower."""
+    from micro_sam.util import training_autocast_dtype
+
+    assert training_autocast_dtype(torch.device(device_type)) is None
 
 
 if __name__ == "__main__":
